pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "fmcore"
dynamic = ["version"]
authors = [
    { name = "Abhishek Divekar", email = "adivekar@utexas.edu" }
]
description = "A specialized toolkit for scaling experimental research with Foundation Models."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license-files = ["LICENSE"]
dependencies = [
    "requests",
    "pyyaml",
    "urllib3",
    "pandas==1.*",
    "numpy==1.*",
    "pydantic==1.10.15",
    "scikit-learn",
    "xlrd",
    "XlsxWriter",
    "openpyxl",
    "fastparquet",
    "pyarrow",
    "s3fs",
    "tqdm",
    "boto3",
    "cloudpickle>=3.0.0",
]

[project.optional-dependencies]
all = [
    "pytest",
    "orjson",
    "ray==2.9.2",
    "ray[default]",
    "ray[tune]",
    "ray[serve]",
    "dask[complete]",
    "dask==2024.2.0",
    "gpustat",
    "nvitop",
    "altair",
    "plotly",
    "plotly-express",
    "bokeh",
    "seaborn",
    "holoviews",
    "hvplot>=0.10.0",
    "matplotlib",
    "tiktoken",
    "torch==2.3.0",
    "transformers==4.42.4",
    "einops",
    "accelerate",
    "deepspeed",
    "sentence_transformers",
    "tokenizers",
    "datasets",
    "bitsandbytes",
    "safetensors",
    "brotli",
    "nltk",
    "evaluate",
    "mauve-text",
    "sentencepiece",
    "tensorboard",
    "aim==3.*",
    "wandb",
    "imageio",
    "faiss-cpu",
]

[tool.hatch.version]
source = "vcs"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 110
fix = true
force-exclude = true
extend-exclude = [
    "__init__.py",
]

[tool.ruff.lint]
fixable = [
    "I",    # Add all rules under isort linter: https://docs.astral.sh/ruff/rules/#isort-i
    "W",    # Add all rules under whitespace: https://docs.astral.sh/ruff/rules/#warning-w
    "E401", # multiple-imports-on-one-line: https://docs.astral.sh/ruff/rules/multiple-imports-on-one-line/
    "E713", # not-in-test: https://docs.astral.sh/ruff/rules/not-in-test/
    "E721", # type-comparison: https://docs.astral.sh/ruff/rules/type-comparison/
    "E722", # bare-except: https://docs.astral.sh/ruff/rules/bare-except/
    "F401", # unused-import: https://docs.astral.sh/ruff/rules/unused-import/
    "F541", # f-string-missing-placeholders: https://docs.astral.sh/ruff/rules/f-string-missing-placeholders/
    "F811", # redefined-while-unused: https://docs.astral.sh/ruff/rules/redefined-while-unused/
    "F841", # unused-variable: https://docs.astral.sh/ruff/rules/unused-variable/
]
ignore = [
    ## Ignored because it makes the code too verbose:
    "E731", # lambda-assignment: https://docs.astral.sh/ruff/rules/lambda-assignment/
    "E741", # ambiguous-variable-name: https://docs.astral.sh/ruff/rules/ambiguous-variable-name/
    
    ## Ignored because of bad interaction with `from typing import *`
    "F405", # undefined-local-with-import-star-usage: https://docs.astral.sh/ruff/rules/undefined-local-with-import-star-usage/
    "F403", # undefined-local-with-import-star: https://docs.astral.sh/ruff/rules/undefined-local-with-import-star/
]